<refactor>[front] refactor comment logic
- Integrate comment server component files to one file
- Place comment tag logic
